pyproject.toml
[tool.poetry]
name = "python-flask-mongoengine"
version = "0.1.1"
description = "A simple and powerful Python+Flask RESTful Template."
authors = ["Francisco de Souza Junior <fsjunior@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9.5"
Flask = "^1.1.2"
flask-smorest = "^0.30.0"
flask-mongoengine = "^1.0.0"
flask-marshmallow = "^0.14.0"
Flask-Cors = "^3.0.9"
gunicorn = "^20.0.4"
marshmallow = "^3.8.0"
python-dotenv = "^0.17.1"
pymongo-migrate = "^0.11.0"
Flask-Caching = "^1.9.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest-flask = "^1.0.0"
pytest-cov = "^2.10.1"
codecov = "^2.1.10"
mypy = "^0.812"
pytest-deadfixtures = "^2.2.1"
pytest-mock = "^3.3.1"
pre-commit = "^2.11.1"
pylint = "^2.7.4"
pylint-mongoengine = "^0.4.0"
Pygments = "^2.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#Lint and checks
[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[mypy]
python_version = 3.9
ignore_missing_imports = true

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = "E203, W503, C0114, C0115, C0116, R0903"

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]